requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
numpy>=1.21.0
pyyaml>=6.0
spacy>=3.5.0
tqdm>=4.65.0
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml[html_clean]>=4.9.0 